[#6] refactor: 통합 테스트를 위한 TestSupport 파일 이동
